Fixed counter of microseconds.

+++ src/Diagnostics/SmoothCounter.cs
@@ -7,10 +7,10 @@
     {
         private int _index = 0;
 
-        private int _totalDeltaTime = 0;
-        private int[] _previousTime;
+        private double _totalDeltaTime = 0;
+        private double[] _previousTime;
 
-        private int _longestTime = 0;
+        private double _longestTime = 0;
 
         private int _totalEntitiesCount = 0;
         private int[] _previousEntityCount;
@@ -20,7 +20,7 @@
         /// <summary>
         /// Average of counter time value over the sample size.
         /// </summary>
-        public int AverageTime => (int)MathF.Round(_totalDeltaTime / (float)_sampleSize);
+        public int AverageTime => (int)MathF.Round((float)(_totalDeltaTime / _sampleSize));
 
         /// <summary>
         /// Average of entities over the sample size.
@@ -30,9 +30,9 @@
         /// <summary>
         /// Maximum value over the sample size.
         /// </summary>
-        public int MaximumTime => _longestTime;
+        public double MaximumTime => _longestTime;
 
-        public SmoothCounter(int size = 1000) => (_sampleSize, _previousTime, _previousEntityCount) = (size, new int[size], new int[size]);
+        public SmoothCounter(int size = 1000) => (_sampleSize, _previousTime, _previousEntityCount) = (size, new double[size], new int[size]);
 
         public void Clear()
         {
@@ -43,11 +43,11 @@
 
             _longestTime = 0;
 
-            _previousTime = new int[_sampleSize];
+            _previousTime = new double[_sampleSize];
             _previousEntityCount = new int[_sampleSize];
         }
 
-        public void Update(int ms, int totalEntities)
+        public void Update(double ms, int totalEntities)
         {
             _index++;
 
@@ -61,7 +61,7 @@
             {
                 _longestTime = ms;
             }
-
+            
             _totalDeltaTime -= _previousTime[_index];
             _totalDeltaTime += ms;
